Create U9 p90-91 Reading
Six sentences have been removed from the article. Choose from the sentences A-G the one which fits each gap 1-6. There is one extra sentence which you do not need to use.

A. These results probably indicate that the changes that were made were long overdue.
B. This isn't to say that everyone considering a drastic change in life should leap into action.
C. They did not enjoy the experience but the results were superb.
D. However, a random change can also lead us into behaving in a way that we might not have imagined.
E. The reason we haven't made them already is probably because we tend to stick with what is safe and familiar.
F. Nonetheless, it is important to know when to stop.
G. He asked people who were agonizing over a life decision to take part in an experiment he had set up.

BIG DECISION AHEAD? JUST TAKE A CHANCE
The decision-making process is never easy. Emotions can cloud our judgement. The more information you have, the easier a decision should be but more choice frequently results in information overload, which can be overwhelming and have tl1e opposite effect. In fact, at times, relying on chance or an outside intervention can actually help us to make better decisions.
The first reason for this is that by deciding to follow a random instruction, we can end up making decisions that we should have been making all along. 
[1] {1:MCS:%0%A~%0%B~%0%C~%0%D~%100%E~%0%F~%0%G}Tossing a coin and deciding that 'if the coin comes up heads, I'll propose or resign or have a baby' may be the only way that some of us can find the energy to make tough decisions.
A few years ago the economist Stephen Levitt found evidence for this theory. 
[2] {2:MCS:%0%A~%0%B~%0%C~%0%D~%0%E~%0%F~%100%G}They were required to log onto the website he had launched, answer some questions and then hand over responsibility to the flip of a digital coin.
As the months went past, Levitt followed up, asking the volunteers whether they had obeyed the coin or ignored it and how they were now feeling about the decision. 
It's an odd experiment but what makes it powerful is that the toss of the coin randomly divided people into two groups, one pushed towards action and one pushed towards caution, depending on whether the coin came up 'heads' or 'tails'. In fact, people did tend to obey the coin, and the majority of those in the 'action' group reported being happier several months on. 
[3] {3:MCS:%100%A~%0%B~%0%C~%0%D~%0%E~%0%F~%0%G}The research also suggests that, if this is the case, we should probably get on with it, with the help of a coin if necessary. Randomness, then, can prompt us into taking actions that we fear. [4] {4:MCS:%0%A~%0%B~%0%C~%100%D~%0%E~%0%F~%0%G}
A simple example of this unpredictable behavior emerged after an Underground strike in London. Three economists looked at data which identified regular commuters who had to find a different route to work during the strike. What was surprising was that one in 20 of these individuals stuck to their new routes after the strike was over. Even commuters, whom we would imagine had perfected the most efficient way to make their daily journey, can come up with a better route when forced to do so.
Many creative types have used this principle to their advantage. Perhaps the most famous example is the Oblique Strategies, a set of cards assembled by artist Peter Schmidt and musician Brian Eno. The cards are full of complicated instructions which were used while working with David Bowie on some of his celebrated albums, and also with the band Coldplay. Great guitarists would find themselves having to play the drums or jump randomly between chords that Eno was indicating on a blackboard. [5] {5:MCS:%0%A~%0%B~%100%C~%0%D~%0%E~%0%F~%0%G}
One can, of course, take the idea too far. It is true that randomness may well help us to think and behave more creatively.
[6] {6:MCS:%0%A~%0%B~%0%C~%0%D~%0%E~%100%F~%0%G}Once we have found a life we love, we are probably better off staying with the status quo.